Explicitly specify is_clang=false for Win MSVC bots
Otherwise they're doing exactly the same as Clang bots.

Also fix warnings that have sneaked in.

TBR=ehmaldonado@webrtc.org

Bug: webrtc:8664

+++ video/picture_id_tests.cc
@@ -278,7 +278,7 @@ class VideoStreamFactory
     // Use the same total bitrates when sending a single stream to avoid
     // lowering the bitrate estimate and requiring a subsequent rampup.
     const int encoder_stream_bps =
-        kEncoderBitrateBps / encoder_config.number_of_streams;
+        kEncoderBitrateBps / static_cast<int>(encoder_config.number_of_streams);
 
     for (size_t i = 0; i < encoder_config.number_of_streams; ++i) {
       streams[i].min_bitrate_bps = encoder_stream_bps;
